payement avec stripe

+++ eventService/src/main/java/com/example/eventService/entities/EventDTO.java
@@ -11,7 +11,7 @@ public class EventDTO {
     private String description;
     private LocalDateTime datedebut;
     private LocalDateTime datefin;
-    private float prix;
+    private double prix;
     private String status;
     private CategoryDTO category;
     private LocalDTO local;
@@ -59,11 +59,11 @@ public class EventDTO {
         this.description = description;
     }
 
-    public float getPrix() {
+    public double getPrix() {
         return prix;
     }
 
-    public void setPrix(float prix) {
+    public void setPrix(double prix) {
         this.prix = prix;
     }
